requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-socketio==5.11.0
sqlalchemy==2.0.35
spotipy==2.24.0
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.0
itsdangerous==2.2.0
python-multipart==0.0.9
